feat(peacewatcher): refacto project and clean build and main
