--- drone_simulator/src/main/scala/Person.scala
@@ -1,4 +1,0 @@
-final case class Person(
-    name: String,
-    score: Double
-)
