+++ peacewatcher_simulator/src/main/scala/peacewatcher/Event.scala
@@ -1,9 +1,13 @@
+package peacewatcher
+
 import java.time.LocalDateTime
 
+final case class Coords(latitude: Double, longitude: Double)
+
 final case class Event(
+    peacewatcher_id : Int,
     timestamp: LocalDateTime,
-    drone_id: Int,
-    location: Tuple2[Double, Double],
+    location: Coords,
     words: List[String],
     persons: List[Person]
 )
