+++ peacewatcher_simulator/src/main/scala/peacewatcher/Person.scala
@@ -1,0 +1,6 @@
+package peacewatcher
+
+final case class Person(
+    name: String,
+    peacescore: Double,
+)
